Refactor HttpRequestContext classes - remove 'BASE_PATH' static variable - split HttpRequest Concrete classes to collection class & item class
