+++ app/src/main/java/com/codesoom/assignment/web/HttpRequestContextBase.java
@@ -11,6 +11,13 @@ public abstract class HttpRequestContextBase {
     private final RequestControllable nullController = httpRequest -> new HttpResponse(HttpStatusCode.NOT_FOUND);
     protected final Map<HttpRequestMethod, RequestControllable> requestControllerMap = new HashMap<>();
     protected final JsonMapper jsonMapper = new JsonMapper();
+    protected final String basePath;
+
+    public abstract boolean matchesPath(String path);
+
+    public HttpRequestContextBase(String basePath) {
+        this.basePath = basePath;
+    }
 
     public void processHttpRequest(HttpRequest httpRequest, HttpExchange httpExchange) throws IOException {
         RequestControllable controller = findController(httpRequest.getMethod());
@@ -29,5 +36,11 @@ public abstract class HttpRequestContextBase {
 
     protected RequestControllable findController(HttpRequestMethod method) {
         return requestControllerMap.getOrDefault(method, nullController);
+    }
+
+    protected long parseIdFromPath(String path) throws NumberFormatException {
+        //path 마지ll막에 '/'이 붙어 있을 것을 대비
+        String idString = path.replace(basePath, "").replaceAll("/", "");
+        return Long.parseLong(idString);
     }
 }
